requirements.txt
# Web Framework
streamlit>=1.39.0

# Azure SDKs
azure-devops==7.1.0b4
azure-search-documents==11.4.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# OpenAI
openai>=1.55.0

# Utilities
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml>=5.3.0

# Data Processing
numpy>=2.0.0
tiktoken>=0.7.0

# Logging and Monitoring
azure-monitor-opentelemetry==1.2.0
